Keylogger
Keylogger Project

A basic keylogger built with Python. It captures keystrokes and logs them to a file. It is intended for educational purposes and not for malicious use.

Features:

Captures keystrokes.
Written in Python using the pynput library.
Disclaimer: For educational purposes only. Unauthorized use may result in legal consequences.